+++ commerce-data/src/main/java/co/hooghly/commerce/repository/TaxRateRepository.java
@@ -22,10 +22,10 @@ public interface TaxRateRepository extends JpaRepository<TaxRate, Long> {
 	TaxRate findOne(Long id);
 	
 	@Query("select t from TaxRate t join fetch t.taxClass join fetch t.merchantStore tm join fetch t.country tc left join fetch t.zone tz left join fetch t.descriptions td left join t.parent where tm.id=?1 AND (tz.id=?2 OR tz IS NULL) and tc.id=?3 and td.language.id=?4 order by t.taxPriority asc")
-	List<TaxRate> findByMerchantAndZoneAndCountryAndLanguage(Integer id, Long zoneId, Integer countryId, Integer languageId);
+	List<TaxRate> findByMerchantAndZoneAndCountryAndLanguage(Integer id, Long zoneId, Long countryId, Integer languageId);
 	
 	@Query("select t from TaxRate t join fetch t.taxClass join fetch t.merchantStore tm join fetch t.country tc left join fetch t.zone tz left join fetch t.descriptions td left join t.parent where tm.id=?1 AND t.stateProvince=?2 and tc.id=?3 and td.language.id=?4 order by t.taxPriority asc")
-	List<TaxRate> findByMerchantAndProvinceAndCountryAndLanguage(Integer id, String province, Integer countryId, Integer languageId);
+	List<TaxRate> findByMerchantAndProvinceAndCountryAndLanguage(Integer id, String province, Long countryId, Integer languageId);
 	
 	
 }
